requirements.txt
autopep8==1.6.0
Django==2.2.17
django-jenkins==0.110.0
djangorestframework==3.13.1
pep8==1.7.1
postgres==3.0.0
psycopg2-binary==2.8.6
psycopg2-pool==1.1
pycodestyle==2.8.0
pyflakes==2.4.0
pytz==2020.4
sqlparse==0.4.2
toml==0.10.2
django-crispy-forms==1.12.0